fix: resolve pipeline quality loss, color compositing, and dynamo runtime errors
This commit implements fixes to resolve VAE fidelity loss and runtime errors within the WearCast AI pipeline.

- VAE Full-Precision Reconstruction:
  - Instantiated the VAE in torch.float32 and explicitly cast garm_tensor to float32 for numerical stability.
  - Added automatic down-casting to float16 for latents returning from prepare_garm_latents and prepare_vton_latents.

- Dynamic Color Compositing Adjustments:
  - Removed strict min/max hard caps on shift_v, replacing them with a statistical shift.
  - Reversed the compositing order of operations so adaptive shadows apply before the final color correction map.

- Torch Dynamo Compilation Fix:
  - Removed the torch._dynamo.config.allow_unspec_int_on_nn_module = True configuration.
  - This defaults torch.compile to static integer dimensions for UNet blocks, resolving the pow_by_natural Sympy exception.

+++ wearcast/pipelines_wearcast/pipeline_wearcast.py
@@ -913,7 +913,7 @@ class WearCastPipeline(DiffusionPipeline, TextualInversionLoaderMixin, LoraLoade
             image_latents = torch.cat([image_latents], dim=0)
 
         # No doubling here; we handle it in the main denoising loop for consistency
-        return image_latents
+        return image_latents.to(dtype=dtype)
     
     def prepare_vton_latents(
         self, image, mask, image_ori, batch_size, num_images_per_prompt, dtype, device, do_classifier_free_guidance, generator=None
@@ -994,9 +994,9 @@ class WearCastPipeline(DiffusionPipeline, TextualInversionLoaderMixin, LoraLoade
 
         # No doubling here; we handle it in the main denoising loop for consistency and to avoid double-doubling
         # Final Device Anchor: Force all return tensors back to GPU
-        image_latents = image_latents.to(device=device)
-        mask = mask.to(device=device)
-        image_ori_latents = image_ori_latents.to(device=device)
+        image_latents = image_latents.to(device=device, dtype=dtype)
+        mask = mask.to(device=device, dtype=dtype)
+        image_ori_latents = image_ori_latents.to(device=device, dtype=dtype)
 
         return image_latents, mask, image_ori_latents
